fix: resolve 34 mypy strict type errors across 13 engine files
- Add explicit type parameters for generic dict/list/Match annotations
- Add typing.Any imports where needed
- Fix incompatible assignment in knowledge_base.py by converting dict_values to list
- Wrap float expression in rate_limiter.py to satisfy no-any-return
- Add typing.cast for compliance.py deepcopy return
- Add str() wrapper for re.Match.group() in i18n.py

+++ src/vibe_pdca/engine/rate_limiter.py
@@ -245,7 +245,7 @@ class BackoffStrategy:
         """
         if attempt > self._max_attempts:
             return -1.0
-        return self._base_delay * (2 ** (attempt - 1))
+        return float(self._base_delay * (2 ** (attempt - 1)))
 
 
 # ── RateLimitDashboard ──
